Compute Mel-filterbank energy features from audio

This function computes Mel-filterbank energy features from audio signals.

get_mel(
  chopped_dir,
  winlen = 0.025,
  winstep = 0.01,
  nfft = 512,
  nfilt = 26,
  lowfreq = 0,
  highfreq = NULL
)

Arguments

chopped_dir

Directory of the individual sound files. Typically created with get_word_sound_files.

winlen

The length of the analysis window in seconds. Default is 0.025.

winstep

The step between successive windows in seconds. Default is 0.01.

nfft

The FFT size. Default is 512.

nfilt

The number of filters in the filterbank. Default is 26.

lowfreq

Lowest band edge of Mel filters. In Hz, default is 0.

highfreq

Highest band edge of Mel filters. In Hz, default is samplerate/2.

Value

A list object.

References

Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2023). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR

James Lyons, Darren Yow-Bang Wang, Gianluca, Hanan Shteingart, Erik Mavrinac, Yash Gaurkar, Watcharapol Watcharawisetkul, Sam Birch, Lu Zhihe, Josef Hölzl, Janis Lesinskis, Henrik Almér, Chris Lord, & Adam Stark. (2020). jameslyons/python_speech_features: release v0.6.1 (0.6.1). Zenodo. https://doi.org/10.5281/zenodo.3607820

Author

Dominic Schmitz

Examples

if (FALSE) mel_features <- get_mel(chopped_dir = "/chopped")